# openapi.yaml
openapi: 3.0.3
info:
  title: Accounting MCP API
  version: "1.0.0"
  description: |
    会計 freee データを共通スキーマで提供する MCP サーバーの API 仕様。
    仕訳帳は `detail` パラメータで `simple` / `standard` / `full` の 3 段階を切り替え可能。
servers:
  - url: http://localhost:3000
paths:
...
  /audit/logs:
    get:
      summary: 監査ログ取得
      parameters:
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - name: event_type
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
        - name: endpoint
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: 監査ログ一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLogEntry'
                  metadata:
                    type: object
  /audit/statistics:
    get:
      summary: 監査統計取得
      parameters:
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        "200":
          description: 監査統計
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditStatisticsResponse'
  /audit/export:
    get:
      summary: 監査ログエクスポート
      parameters:
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - name: event_type
          in: query
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        "200":
          description: エクスポートされたログ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLogEntry'
            text/csv:
              schema:
                type: string
  /audit/verify/{id}:
    get:
      summary: 監査ログ整合性検証
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 整合性検証結果
          content:
            application/json:
              schema:
                type: object
                properties:
                  log_id:
                    type: integer
                  integrity_valid:
                    type: boolean
                  verified_at:
                    type: string
                    format: date-time
  /audit/summary/events:
    get:
      summary: イベントタイプ別サマリ
      parameters:
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        "200":
          description: イベント別サマリ
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: object
  /audit/summary/users:
    get:
      summary: ユーザー別サマリ
      parameters:
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        "200":
          description: ユーザー別サマリ
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: object
components:
  parameters:
...
    StartDate:
      name: start_date
      in: query
      schema:
        type: string
        format: date
    EndDate:
      name: end_date
      in: query
      schema:
        type: string
        format: date
  schemas:
...
    AuditLogEntry:
      type: object
      properties:
        id:
          type: integer
        timestamp:
          type: string
          format: date-time
        event_type:
          type: string
        user_id:
          type: string
        method:
          type: string
        endpoint:
          type: string
        response_status:
          type: integer
    AuditStatisticsResponse:
      type: object
      properties:
        statistics:
          type: object
          properties:
            summaries:
              type: array
              items:
                type: object
                properties:
                  date:
                    type: string
                  total_requests:
                    type: integer
                  unique_users:
                    type: integer
            totals:
              type: object